pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = ""
authors = [
    {name = "wangchenyu",email = "1542169671@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "grpcio (>=1.76.0,<2.0.0)",
    "grpcio-tools (>=1.76.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "numpy (<2.3)",
    "soundfile (>=0.13.1,<0.14.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "typer (>=0.20.0,<0.21.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "mcp (>=1.26.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "nexus", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.2.0"
pytest-cov = "^7.0.0"

# [tool.poetry.scripts]
# speecharm = "speecharm.commands.app:app"

[tool.poetry.scripts]
nexus = "nexus.commands.app:app"



[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
